project/requirements.txt
requests>=2.31.0,<3.0
psycopg2>=2.8.6,<2.9
mysqlclient==2.1.0
Django==3.2.5
celery>5.2.3,<6.0
redis==3.5.3
django-celery-beat==2.5.0
djangorestframework>=3.12.4,<3.13
psycopg2>=2.8.6,<2.9
drf-spectacular>=0.15.1,<0.16
tweepy==4.14.0